requirements.txt
anyio==4.3.0
click==8.1.7
graphql-core==3.2.3
h11==0.14.0
idna==3.7
libcst==1.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.225.1
typer==0.12.3
typing_extensions==4.11.0
uvicorn==0.29.0
fastapi==0.110.1
